[android] Do not accidentally color navbar before API 26, fix #1913
We cannot change navbar buttons color before API 26 but we changed
navbar color starting from API 23 because of the styles resource (we
defined it there so that system shows the right navbar color for splash
screen). This commit removes navbar color from values-v23 and defines
another style resource for API 26.